Update README.md
This application collects information from clients by having them fill out forms with their information. A consumer who has registered on the internet has the option of booking the car that he or she requires. The suggested system is an online, fully integrated system. It effectively and efficiently automates manual procedures. This automatic system assists customers by filling up the data based on their specifications. It covers the sort of car they 
are looking to hire as well as the location. The goal of this system is to create a website 
where individuals from all across the state can book their vehicles and requirements. Car 
rental firms make a profit by owning or leasing a large fleet of automobiles and lending 
them to consumers for a price. Rental fleets can be framed in a variety of ways: They can 
be purchased outright (known as 'risk vehicles' because the car rental company is betting 
on how much the vehicle will sell for when it is removed from service), leased, or owned 
under a guaranteed buy-back system arranged directly through a manufacturer or 
manufacturer's financial arm (known as repurchase vehicles' because the producer 
outlays the money).